server/lambda/requirements.txt
tensorflow-cpu==2.6.0
Pillow==8.4.0
